Fixed 64bit related compiler warnings that cause build to fail due to being treated as errors.

+++ src/avdecc/mcDomainManager.cpp
@@ -79,7 +79,7 @@ public:
 			});
 
 		connect(&_sequentialAcmpCommandExecuter, &commandChain::SequentialAsyncCommandExecuter::progressUpdate, this,
-			[this](uint32_t const completedCommands, uint32_t const totalCommands)
+			[this](size_t const completedCommands, size_t const totalCommands)
 			{
 				emit applyMediaClockDomainModelProgressUpdate(roundf(((float)completedCommands) / totalCommands * 100));
 			});
